backend/requirements.txt
fastapi
uvicorn
requests
openai
python-dotenv
jinja2
aiofiles